Refactor Apollo client initialization
