+++ src/index.ts
@@ -1,42 +1,8 @@
-import { ApolloCache } from "apollo-cache";
 import { ApolloClient } from "apollo-client";
-import { ApolloLink } from "apollo-link";
-import { BatchHttpLink } from "apollo-link-batch-http";
-import { RetryLink } from "apollo-link-retry";
 
 import { SaleorAPI } from "./api";
 import { ConfigInput } from "./types";
 import APIProxy from "./api/APIProxy";
-import { authLink, invalidTokenLinkWithTokenHandler } from "./auth";
-
-export function createSaleorClient(
-  cache: ApolloCache<any>,
-  links: ApolloLink[]
-) {
-  return new ApolloClient({
-    cache,
-    link: ApolloLink.from(links),
-  });
-}
-
-export const createSaleorLinks = ({
-  apiUrl,
-  tokenExpirationCallback,
-}: {
-  apiUrl: string;
-  tokenExpirationCallback: () => void;
-}) => {
-  const invalidTokenLink = invalidTokenLinkWithTokenHandler(
-    tokenExpirationCallback
-  );
-
-  return [
-    invalidTokenLink,
-    authLink,
-    new RetryLink(),
-    new BatchHttpLink({ uri: apiUrl }),
-  ];
-};
 
 export class SaleorManager {
   private apiProxy: APIProxy;
@@ -72,6 +38,9 @@ export class SaleorManager {
 }
 
 export * from "./auth";
+export * from "./cache";
+export * from "./links";
+export * from "./client";
 export * from "./gqlTypes/globalTypes";
 
 // FIXME: It's imported here because it's not a monorepo yet
